test: move tests for message validation
